🚀 chore(package.json): add new dependencies and devDependencies This commit adds two new dependencies to the project: `import-meta-resolve` and `acorn`. The former is used to resolve import meta properties in ESM modules, while the latter is a dependency of `esmutils`. Additionally, a new devDependency `jsonc-parser` is added to the `pkg-def` package.
🚀 chore(.eslintignore, .gitignore, package.json): add new files and directories
This commit adds new files and directories to the project. The `.eslintignore` file is added to the `esmutils` package to ignore certain directories and files from being linted. The `.gitignore` file is also added to the `esmutils` package to ignore certain directories and files from being tracked by git. The `package.json` files are added to the `esmutils` package's test fixtures to simulate different types of modules being